Трей стал намного лучше: Изменение громкости через скролл на винде, собственный дизайн окна. Больше удобств. Настройка оптимизации окна, TODO добавить настройку

+++ Core/Audio/AudioPlayer.cs
@@ -863,7 +863,7 @@ public sealed class AudioPlayer : IAsyncDisposable, IDisposable
         {
             Sources.CachingStreamSource caching => caching.DownloadedBytes,
             Sources.LocalFileSource => pipeline.Source.IsFullyBuffered
-                ? (long)(pipeline.StreamInfo.DurationMs * pipeline.StreamInfo.Bitrate / 8)
+                ? pipeline.StreamInfo.DurationMs * pipeline.StreamInfo.Bitrate / 8
                 : 0,
             _ => (long)(pipeline.Source.BufferProgress / 100.0
                 * pipeline.StreamInfo.DurationMs * pipeline.StreamInfo.Bitrate / 8)
